splittet up member table
